alerts with swal

+++ src/components/tasks/Tasks.js
@@ -3,6 +3,8 @@ import { todoService } from '../../services/todoService';
 import Form from '../form/Form';
 import Task from '../task/Task';
 import { CtTasks } from './Tasks.styled';
+import Swal from 'sweetalert2';
+import withReactContent from 'sweetalert2-react-content';
 
 const api = todoService();
 
@@ -10,6 +12,8 @@ function Tasks() {
 
     const [tasks, setTasks] = useState([]);
     const [editTask, setEditTask] = useState();
+
+    const MySwal = withReactContent(Swal);
 
     useEffect(() => {  
         api.get().then((res) => setTasks(res.data));
@@ -21,7 +25,14 @@ function Tasks() {
                 setTasks([...tasks, res.data]);
             };
         });
-        alert('Su tarea ha sido creada correctamente');
+        ;
+        MySwal.fire({
+            position: 'top-end',
+            icon: 'success',
+            title: 'Tu tarea ha sido creada correctamente',
+            showConfirmButton: false,
+            timer: 1500
+        });
     };
 
     const updateTask = (data) => {
@@ -31,17 +42,37 @@ function Tasks() {
                 setTasks(editedTask);
             };
         });
-        alert('Su tarea ha sido editada correctamente');
+        MySwal.fire({
+            position: 'top-end',
+            icon: 'success',
+            title: 'Tu tarea ha sido editada correctamente',
+            showConfirmButton: false,
+            timer: 1500
+        });
     };
 
     const deleteTask = (id) => {
-        api.destroy(id).then((res) => {
-            if (res.status === 200) {
-                const remainTasks = tasks.filter(task => task.id !== id);
-                setTasks(remainTasks);
-            };
-        });
-        alert('Su tarea ha sido borrada correctamente');
+        Swal.fire({
+            title: '¿Quieres borrar esta tarea?',
+            text: "¡No podrás revertir esta acción!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: 'var(--main-color)',
+            cancelButtonColor: 'var(--warning-color)',
+            confirmButtonText: '¡Si, bórrala!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                api.destroy(id).then((res) => {
+                    const remainTasks = tasks.filter(task => task.id !== id);
+                    setTasks(remainTasks);
+                });
+                Swal.fire(
+                    '¡Borrada!',
+                    'Tu tarea ha sido borrada',
+                    'success'
+                )
+            }
+        })
     };
  
   return (
